==> 1. data_pipeline.ipynb <==
# import library
import pandas as pd 
import joblib

from sklearn.model_selection import train_test_split

# read data function
def read_data(file_path_name, ind_col):
  data = pd.read_csv(file_path_name, index_col=ind_col)
  print('shape :', data.shape)
  return data

# show shape of data
data = read_data('data/raw/data.csv', 0)

# show the five rows of data
data.head()

# data shape
data.shape

# statistical summary of data
data.describe()

# split input output function
def split_input_output(dataset, target_var):
  y = dataset[target_var]
  X = data.drop(target_var, axis=1)
  
  print('shape X:', X.shape)
  print('shape y:', y.shape)
  
  joblib.dump(X, 'data/output/input.pkl')
  joblib.dump(y, 'data/output/output.pkl')

# define target variable
TARGET_VAR = 'SeriousDlqin2yrs'

# running the split input output function
split_input_output(data, TARGET_VAR)

# split train test function
def split_train_test(data_x, data_y, size_test, size_valid):
  X_train, X_test, y_train, y_test = train_test_split(data_x, data_y, 
                                                      test_size = size_test, random_state=42)
  X_train, X_valid, y_train, y_valid = train_test_split(X_train, y_train, 
                                                        test_size = size_valid, random_state=42)
  
  print('X_train shape :', X_train.shape)
  print('y_train shape :', y_train.shape)
  print('X_valid shape :', X_valid.shape)
  print('y_valid shape :', y_valid.shape)
  print('X_test shape :', X_test.shape)
  print('y_test shape :', y_test.shape)
  
  joblib.dump(X_train, 'data/output/X_train.pkl')
  joblib.dump(y_train, 'data/output/y_train.pkl')
  joblib.dump(X_valid, 'data/output/X_valid.pkl')
  joblib.dump(y_valid, 'data/output/y_valid.pkl')
  joblib.dump(X_test, 'data/output/X_test.pkl')
  joblib.dump(y_test, 'data/output/y_test.pkl')

# load X and y data 
X = joblib.load('data/output/input.pkl')
y = joblib.load('data/output/output.pkl')

# running the split train test function
split_train_test(X, y, 0.2, 0.2)
