==> 3. data_preprocessing.ipynb <==
# import library
import pandas as pd
import numpy as np
import joblib

from sklearn.preprocessing import StandardScaler
from sklearn.impute import SimpleImputer

from imblearn.under_sampling import RandomUnderSampler

# Plan preprocessing

**Summary EDA**:
- Data yang `NumberOfTimes90DaysLate == 96` atau `NumberOfTimes90DaysLate == 98` di delete
- Data yang `RevolvingUtilizationOfUnsecuredLines > 1.35` di delete
- Missing value imputation 
  - `MonthlyIncome` : Median
  - `NumberOfDependents` : 0.0

## Load dataset

# load the dataset function
def load_dataset():
  # load train data
  X_train = joblib.load('data/output/X_train.pkl')
  y_train = joblib.load('data/output/y_train.pkl')
  
  # load valid data
  X_valid = joblib.load('data/output/X_valid.pkl')
  y_valid = joblib.load('data/output/y_valid.pkl')
  
  # load test data
  X_test = joblib.load('data/output/X_test.pkl')
  y_test = joblib.load('data/output/y_test.pkl')
  
  # print shape
  print('X_train shape :', X_train.shape)
  print('y_train shape :', y_train.shape)
  print('X_valid shape :', X_valid.shape)
  print('y_valid shape :', y_valid.shape)
  print('X_test shape :', X_test.shape)
  print('y_test shape :', y_test.shape)
  return X_train, X_valid, X_test, y_train, y_valid, y_test

# running load dataset function
X_train, X_valid, X_test, y_train, y_valid, y_test = load_dataset()

## Preprocess Train

1. Drop values in `NumberOfTimes90DaysLate` bigger than or equal to 96

# clean late data function
def clean_late_data(X, y):
  """Function to clean NumberOfTimes90DaysLate columns"""
  # find index drop
  drop_condition = X['NumberOfTimes90DaysLate'] >= 96
  index_to_drop = X[drop_condition].index.tolist()
  
  # drop data
  X_drop = X.drop(index = index_to_drop)
  y_drop = y.drop(index = index_to_drop)
  
  # print 
  print('X shape :', X_drop.shape)
  print('y shape :', y_drop.shape)
  
  return X_drop, y_drop

# running the claen date data in data train
X_train_drop_1, y_train_drop_1 = clean_late_data(X_train, y_train)

2. Data yang `RevolvingUtilizationOfUnsecuredLines > 1.35` di delete

# clean unsecured data function
def clean_unsecured_data(X, y):
  """Function to clean RevolvingUtilizationOfUnsecuredLines columns from outlier"""
  q1, q3 = np.quantile(X['RevolvingUtilizationOfUnsecuredLines'], q = [0.25, 0.75])
  iqr = q3 - q1
  ub = q3 + 1.5*iqr
  lb = q1 - 1.5*iqr
  
  # filter data
  drop_condition_1 = X['RevolvingUtilizationOfUnsecuredLines'] > ub
  drop_condition_2 = X['RevolvingUtilizationOfUnsecuredLines'] < lb
  index_to_drop = X[drop_condition_1 | drop_condition_2].index.tolist()
  
  # drop data 
  X_drop = X.drop(index = index_to_drop)
  y_drop = y.drop(index = index_to_drop)
  
  # print
  print('X shape :', X_drop.shape)
  print('y shape :', y_drop.shape)
  
  return X_drop, y_drop

# running the clean unsecured data function
X_train_drop_2, y_train_drop_2 = clean_unsecured_data(X_train_drop_1, y_train_drop_1)

3. Missing value imputation 
  - `MonthlyIncome` : Median
  - `NumberOfDependents` : 0.0

# fit imputer function
def fit_imputer(data):
  """Function to fit imputer (constant & median)"""
  # create imputer zero and median
  constant_imputer = SimpleImputer(missing_values = np.nan,
                                   strategy= 'constant',
                                   fill_value= 0.0)
  
  median_imputer = SimpleImputer(missing_values = np.nan,
                                 strategy = 'median')
  
  # fit imputer
  constant_imputer.fit(data[['NumberOfDependents']])
  median_imputer.fit(data[['MonthlyIncome']])
  
  # dump imputer 
  joblib.dump(constant_imputer, 'data/output/constant_imputer.pkl')
  joblib.dump(median_imputer, 'data/output/median_imputer.pkl')
  
  return constant_imputer, median_imputer

# running the fit imputer function
constant_imputer, median_imputer = fit_imputer(data = X_train_drop_2)

# show number of missing value in X_train_drop_2
X_train_drop_2.isna().sum()

# transform imputer function
def transform_imputer(data, constant_imputer, median_imputer):
  """Function to transform imputer"""
  data = data.copy()
  
  # build impute constant and median variable
  impute_constant = constant_imputer.transform(data[['NumberOfDependents']])
  impute_median = median_imputer.transform(data[['MonthlyIncome']])
  
  # applied impute constant and impute median in NumberOfDependents and MonthlyIncome 
  data['NumberOfDependents'] = impute_constant
  data['MonthlyIncome'] = impute_median
  
  # print shape of data
  print('data shape :', data.shape)
  
  return data

# running the tranform imputer function
X_train_imputed = transform_imputer(data=X_train_drop_2,
                                    constant_imputer = constant_imputer,
                                    median_imputer = median_imputer)

# data train don't have missing values
X_train_imputed.isna().sum()

4. Standardize data

# standardize function
def fit_standardize(data):
  """Find standardizer data"""
  standardizer = StandardScaler()
  
  # fit standardizer
  standardizer.fit(data)
  
  # dump standardizer 
  joblib.dump(standardizer, 'data/output/standardizer.pkl')
  
  return standardizer

# fit standardizer 
standardizer = fit_standardize(data=X_train_imputed)

# show standardizer
standardizer

# transform standardize function
def transform_standardize(data, standardizer):
  """Function to standardizer data"""
  data_standard = pd.DataFrame(standardizer.transform(data))
  data_standard.columns = data.columns
  data_standard.index = data.index
  return data_standard

# Transform standardize
X_train_std = transform_standardize(data = X_train_imputed,
                                     standardizer = standardizer)

# show X_train_std 
X_train_std

# statistical summary in X_train_std
X_train_std.describe()

Balancing data

# check balancing data
y_train_drop_2.value_counts(normalize=True)

Resampling using downsampling (only for training data)

# random undersampler function
def random_undersampler(X, y):
  """Function to under sample the majority data"""
  # Create resampling object
  ros = RandomUnderSampler(random_state = 42)
  
  # Balancing the set data
  X_resample, y_resample = ros.fit_resample(X, y)
  
  # print
  print('Distribution before resampling :')
  print(y.value_counts())
  print('Distribution after resampling :')
  print(y_resample.value_counts())
  
  return X_resample, y_resample

# running random undersampler function
X_train_clean, y_train_clean = random_undersampler(X_train_std, y_train_drop_2)

Drop all preprocessor

preprocessor = {
  'constant_imputer' : constant_imputer,
  'median_imputer' : median_imputer,
  'standardizer' : standardizer
}

joblib.dump(preprocessor, 'data/output/preprocessor.pkl')

# Preprocess all

# clean data function
def clean_data(data, constant_imputer, median_imputer, standardizer):
  """Function to clean data"""
  # impute missing value
  data_imputed = transform_imputer(data, constant_imputer, median_imputer)
  
  # standardize data
  data_standard = transform_standardize(data_imputed, standardizer)
  
  return data_standard

# _preprocess_data function
def _preprocess_data(data):
  """Function to preprocess data"""
  # load preprocess
  preprocessor = joblib.load('data/output/preprocessor.pkl')
  constant_imputer = preprocessor['constant_imputer']
  median_imputer = preprocessor['median_imputer']
  standardizer = preprocessor['standardizer']
  
  data_clean = clean_data(data,
                          constant_imputer,
                          median_imputer,
                          standardizer)
  
  return data_clean

## preprocess all

Generate preprocessor

# generate preprocessor function
def generate_preprocessor():
  """Function to generate preprocessor"""
  # load data 
  X = joblib.load('data/output/X_train.pkl')
  y = joblib.load('data/output/y_train.pkl')
  
  # drop unusual data
  X, y = clean_late_data(X, y)
  X, y = clean_unsecured_data(X, y)
  
  # Generate preprocessor: imputer
  constant_imputer, median_imputer = fit_imputer(data = X)
  X_imputed = transform_imputer(X, constant_imputer, median_imputer)
  
  # Generate preprocessor: standardizer
  standardizer = fit_standardize(X_imputed)
  
  # dump file = 
  preprocessor = {
    'constant_imputer' : constant_imputer,
    'median_imputer' : median_imputer,
    'standardizer' : standardizer
  }
  
  joblib.dump(preprocessor, 'data/output/preprocessor.pkl')
  
  return preprocessor

# running generate preprocessor function
preprocessor = generate_preprocessor()

# for train

# preprocess data function
def preprocess_data(type='train'):
  """Function to preprocess train data"""
  # load data
  X = joblib.load(f'data/output/X_{type}.pkl')
  y = joblib.load(f'data/output/y_{type}.pkl')
  
  # drop unusual data
  X, y = clean_late_data(X, y)
  X, y = clean_unsecured_data(X, y)
  
  # preprocess data = 
  X_clean = _preprocess_data(X)
  y_clean = y
  
  # Undersampling
  X_clean, y_clean = random_undersampler(X_clean, y_clean)
  
  # print shape
  print('X clean shape :', X_clean.shape)
  print('y clean shape :', y_clean.shape)
  
  joblib.dump(X_clean, f'data/output/X_{type}_clean.pkl')
  joblib.dump(y_clean, f'data/output/y_{type}_clean.pkl')
  
  return X_clean, y_clean

# transform X_train
X_train_clean, y_train_clean = preprocess_data(type='train')

# transform X_valid
X_valid_clean, y_valid_clean = preprocess_data(type='valid')

# transform X_test
X_test_clean, y_test_clean = preprocess_data(type='test')

# show list of X_train_clean column
X_train_clean.columns.tolist()
