==> 4. modeling.ipynb <==
# import library
import pandas as pd 
import numpy as np
import joblib
import copy

from sklearn.neighbors import KNeighborsClassifier
from sklearn.linear_model import LogisticRegression
from xgboost import XGBClassifier

from sklearn.model_selection import RandomizedSearchCV
from sklearn.metrics import roc_auc_score

# Create Model

Model :
- KNN
- Logistic Regression
- XGBoost

Define params

# create model param function
def create_model_param():
  """Create the model objedts"""
  knn_params = {
    'n_neighbors' : [50, 100, 200],
  }
  
  lgr_params = {
    'penalty' : ['l1', 'l2'],
    'C' : [0.01, 0.1],
    'max_iter' : [100, 300, 500]
  }
  
  xgb_params = {
    'n_estimators' : [5, 10, 25, 50]
  }
  
  # create model params
  list_of_param = {
    'KNeighborsClassifier' : knn_params,
    'LogisticRegression' : lgr_params,
    'XGBClassifier' : xgb_params
  }
  
  return list_of_param

Define models

# create model object function
def create_model_object():
  """Create the model objects"""
  print('Creating model objects')
  
  # create model objects
  knn = KNeighborsClassifier()
  lgr = LogisticRegression(solver='liblinear')
  xgb = XGBClassifier()
  
  # create list of model
  list_of_model = [
    {'model_name' : knn.__class__.__name__, 'model_object' : knn},
    {'model_name' : lgr.__class__.__name__, 'model_object' : lgr},
    {'model_name' : xgb.__class__.__name__, 'model_object' : xgb}
  ]
  
  return list_of_model

Do the cross validation

# train model function
def train_model():
  """Function to get the best model"""
  # load dataset
  X_train = joblib.load('data/output/X_train_clean.pkl')
  y_train = joblib.load('data/output/y_train_clean.pkl')
  X_valid = joblib.load('data/output/X_valid_clean.pkl')
  y_valid = joblib.load('data/output/y_valid_clean.pkl')
  
  # create list of params & models
  list_of_param = create_model_param()
  list_of_model = create_model_object()
  
  # list of trained model
  list_of_tuned_model = {}
  
  # train model
  for base_model in list_of_model:
    # current condition
    model_name = base_model['model_name']
    model_obj = copy.deepcopy(base_model['model_object'])
    model_param = list_of_param[model_name]
    
    # debug message
    print('Training model :', model_name)
    
    # create model object
    model = RandomizedSearchCV(estimator=model_obj,
                               param_distributions=model_param,
                               n_iter=5,
                               cv=5,
                               random_state=123,
                               n_jobs=1,
                               verbose=10,
                               scoring='roc_auc')
    
    # train model
    model.fit(X_train, y_train)
    
    # predict
    y_pred_proba_train = model.predict_proba(X_train)[:, 1]
    y_pred_proba_valid = model.predict_proba(X_valid)[:, 1]
    
    # get score
    train_score = roc_auc_score(y_train, y_pred_proba_train)
    valid_score = roc_auc_score(y_valid, y_pred_proba_valid)
    
    # append 
    list_of_tuned_model[model_name] = {
      'model': model,
      'train_auc': train_score,
      'valid_auc': valid_score,
      'best_params': model.best_params_
    }
    
    print('Done training')
    print('')
  
  joblib.dump(list_of_param, 'data/model/list_of_param.pkl')
  joblib.dump(list_of_model, 'data/model/list_of_model.pkl')
  joblib.dump(list_of_tuned_model, 'data/model/list_of_tuned_model.pkl')
  
  return list_of_param, list_of_model, list_of_tuned_model
  
    

# running the train model function
list_of_param, list_of_model, list_of_tuned_model = train_model()

# show list of tuned model
list_of_tuned_model

Get the best model

# get best model function
def get_best_model():
  """Function to get the best model"""
  # load tuned model
  list_of_tuned_model = joblib.load('data/model/list_of_tuned_model.pkl')
  
  # get the best model
  best_model_name = None
  best_model = None
  best_performance = -99999
  best_model_param = None
  
  for model_name, model in list_of_tuned_model.items():
    if model['valid_auc'] > best_performance:
      best_model_name = model_name
      best_model = model['model']
      best_performance = model['valid_auc']
      best_model_param = model['best_params']
      
  # save the best model
  joblib.dump(best_model, 'data/model/best_model.pkl')
  
  # print
  print('=============================================')
  print('Best model         :', best_model_name)
  print('Metric score       :', best_performance)
  print('Best model params  :', best_model_param)
  print('=============================================')
  
  return best_model
  

# show the result of the best model and parameters
best_model = get_best_model()
